+++ src/Event/Dispatcher.php
@@ -46,9 +46,7 @@ class Dispatcher implements EventDispatcherInterface
      */
     public function addListener($eventName, $listener, $priority = 0)
     {
-        $this->events[] = $eventName;
-
-        $this->dispatcher->listen($eventName, $listener, $priority);
+        throw new \Exception('Please use `Event::listen()`.');
     }
 
     /**
@@ -56,17 +54,7 @@ class Dispatcher implements EventDispatcherInterface
      */
     public function addSubscriber(EventSubscriberInterface $subscriber)
     {
-        foreach ($subscriber->getSubscribedEvents() as $eventName => $params) {
-            if (is_string($params)) {
-                $this->addListener($eventName, [$subscriber, $params]);
-            } elseif (is_string($params[0])) {
-                $this->addListener($eventName, [$subscriber, $params[0]], isset($params[1]) ? $params[1] : 0);
-            } else {
-                foreach ($params as $listener) {
-                    $this->addListener($eventName, [$subscriber, $listener[0]], isset($listener[1]) ? $listener[1] : 0);
-                }
-            }
-        }
+        throw new \Exception('Please use `Event::subscribe()`.');
     }
 
     /**
@@ -74,15 +62,7 @@ class Dispatcher implements EventDispatcherInterface
      */
     public function removeListener($eventName, $listener)
     {
-        $listeners = $this->getListeners($eventName);
-
-        $this->dispatcher->forget($eventName);
-
-        foreach ($listeners as $l) {
-            if ($l !== $listener) {
-                $this->addListener($eventName, $l);
-            }
-        }
+        throw new \Exception('Please use `Event::forget()`.');
     }
 
     /**
@@ -90,15 +70,7 @@ class Dispatcher implements EventDispatcherInterface
      */
     public function removeSubscriber(EventSubscriberInterface $subscriber)
     {
-        foreach ($subscriber->getSubscribedEvents() as $eventName => $params) {
-            if (is_array($params) && is_array($params[0])) {
-                foreach ($params as $listener) {
-                    $this->removeListener($eventName, [$subscriber, $listener[0]]);
-                }
-            } else {
-                $this->removeListener($eventName, [$subscriber, is_string($params) ? $params : $params[0]]);
-            }
-        }
+        throw new \Exception('Removing event subscribers is not supported');
     }
 
     /**
@@ -106,17 +78,7 @@ class Dispatcher implements EventDispatcherInterface
      */
     public function getListeners($eventName = null)
     {
-        if (! is_null($eventName)) {
-            return $this->dispatcher->getListeners($eventName);
-        }
-
-        $sorted = [];
-
-        foreach ($this->events as $eventName) {
-            $sorted = array_merge($sorted, $this->dispatcher->getListeners($eventName));
-        }
-
-        return array_filter($sorted);
+        throw new \Exception('Please use `Event::getListeners()`.');
     }
 
     /**
@@ -124,7 +86,7 @@ class Dispatcher implements EventDispatcherInterface
      */
     public function getListenerPriority($eventName, $listener)
     {
-        return array_search($listener, $this->getListeners($eventName), true);
+        throw new \Exception('Event priority is not supported anymore in Laravel.');
     }
 
     /**
@@ -132,6 +94,6 @@ class Dispatcher implements EventDispatcherInterface
      */
     public function hasListeners($eventName = null)
     {
-        return (bool) count($this->getListeners($eventName));
+        throw new \Exception('Please use `Event::hasListeners()`.');
     }
 }