+++ tests/Event/Subscriber.php
@@ -11,22 +11,6 @@ class Subscriber implements EventSubscriberInterface
      */
     public static function getSubscribedEvents()
     {
-        return [
-            'foo' => 'fooListener',
-            'bar' => ['barListener', 1],
-            'baz' => [['bazListener', 2]],
-        ];
-    }
-
-    public function fooListener()
-    {
-    }
-
-    public function barListener()
-    {
-    }
-
-    public function bazListener()
-    {
+        return [];
     }
 }