+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All Notable changes to `laravel-state-machine` will be documented in this file
 
-## 1.0.0 - 201X-XX-XX
+## 1.2.0 - 2017-08-27
+
+- Simplify event dispatcher implementation
+
+## 1.1.1 - 2017-01-31
+
+- Add support for Laravel 5.4
+
+## 1.1.0 - 2017-01-19
+
+- Implement event dispatcher
+
+## 1.0.0 - 2017-01-14
 
 - initial release